fetch data of person, #222
