+++ client/routes/person.js
@@ -1,42 +1,32 @@
-var fetch = require('fetch-ponyfill')();
 var Templates = require('../templates');
-var JSONToHTML = require('../../lib/transforms/json-to-html-data');
 var initJqueryComp = require('../lib/init-jquery-components.js');
+var getDataItem = require('../lib/get-data-item');
 
 module.exports = function (page) {
-  page('/people/:id', enter);
+  page('/people/:id', load, render);
+};
 
-  function enter (ctx) {
-    var pageEl = document.getElementsByTagName('main')[0];
-
-    var id = ctx.params.id;
-    var url = '/people/' + id;
-
+function load (ctx, next) {
+  if (!ctx.isInitialRender) {
     var opts = {
       headers: { Accept: 'application/vnd.api+json' }
     };
-
-    fetch(url, opts)
-    .then(function (res) {
-      if (res.ok) {
-        return res.json();
-      } else {
-        return Promise.reject(new Error(res.status + ' Failed to fetch results'));
-      }
-    })
-    .then(function (json) {
-      if (json.errors) return Promise.reject(json.errors[0]);
-      var data = JSONToHTML(json);
-      data.page = 'person';
-      pageEl.innerHTML = Templates['people'](data);
-      document.getElementsByTagName('title')[0].textContent = data.titlePage;
-      window.scrollTo(0, 0);
-    })
-    .then(function () {
-      initJqueryComp();
-    })
-    .catch(function (err) {
-      console.error('Failed to find person', err);
+    var id = ctx.params.id;
+    var url = '/people/' + id;
+    getDataItem(url, opts, function (data) {
+      ctx.state.data = data;
+      next();
     });
   }
-};
+}
+
+function render (ctx, next) {
+  var pageEl = document.getElementsByTagName('main')[0];
+  ctx.state.data.page = 'person';
+  pageEl.innerHTML = Templates['people'](ctx.state.data);
+  document.getElementsByTagName('title')[0].textContent = ctx.state.data.titlePage;
+  window.scrollTo(0, 0);
+  initJqueryComp();
+}
+
+// TODO add listeners for the search input
